+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 + Added blocks/second display to crop milestone gui and made all crop milestone gui elements customizable/toggleable.
 + Added farming armor drops counter.
 + Added **Colored Name** - Show the visitor name in the color of the rarity.
++ Added **Visitor Item Preview** - Show the base type for the required items next to new visitors (Note that some visitors may require any crop)
 
 
 ### Features from other Mods